- Added user check to prevent not authorized users from editing or deleting blogs, or posts.

+++ blog/views.py
@@ -1,7 +1,9 @@
 from django.shortcuts import render, redirect, get_object_or_404
 from django.contrib.auth.decorators import login_required
+from django.http import Http404
 from .models import Blog, BlogPost, Like
 from .forms import BlogForm, BlogPostForm
+
 
 
 def home(request):
@@ -99,6 +101,9 @@ def edit_blog(request, blog_id):
     # Fetch blog to be edited
     blog = get_object_or_404(Blog, id=blog_id)
 
+    if blog.author != request.user:
+        raise Http404
+
     if request.method != "POST":
         form = BlogForm(instance=blog)
     else:
@@ -118,6 +123,10 @@ def delete_blog(request, blog_id):
     """View for deleting a blog."""
     # Fetch blog in question
     blog = get_object_or_404(Blog, id=blog_id)
+
+    if blog.author != request.user:
+        raise Http404
+
     # If the user has pressed the confirmation button
     if request.method == "POST":
         blog.delete()
@@ -133,6 +142,9 @@ def edit_post(request, post_id):
     post = get_object_or_404(BlogPost, id=post_id)
 
     blog = post.blog
+    if blog.author != request.user:
+        raise Http404
+
     if request.method != 'POST':
         form = BlogPostForm(instance=post)
     else:
@@ -154,6 +166,9 @@ def delete_post(request, post_id):
     """View for deleting a post"""
     post = get_object_or_404(BlogPost, id=post_id)
     blog = post.blog
+
+    if blog.author != request.user:
+        raise Http404
 
     if request.method == 'POST':
         post.delete()
